Dynamic Array Problem
Declare a 2-dimensional array, arr, of  empty arrays. All arrays are zero indexed.
Declare an integer, lastAnswer, and initialize it to 0.

There are 2 types of queries, pa given as an array of strings for you to parse:

1. Query: 1 x y
   1. Let idx = ((x ⊕ lastAnswer) % n).
   2. Append the integer y to arr[idx].
2. Query: 2 x y
   1. Let idx = ((x ⊕ lastAnswer) % n).
   2. Assign the value arr[idx][y%size(arr[idx])] to lastAnswer.
   3. Store the new value of lastAnswer to an answers array.

Note: ⊕ is the bitwise XOR operation, which corresponds to the ^ operator in most languages. Learn more about it on Wikipedia.% is the modulo operator.
Finally, size(arr[idx]) is the number of elements in arr[idx]

Function Description
Complete the dynamicArray function below.
dynamicArray has the following parameters:
- int n: the number of empty arrays to initialize in arr
- string queries[q]: query strings that contain 3 space-separated integers

Returns
int[]: the results of each type 2 query in the order they are presented

Input Format
The first line contains two space-separated integers, n, the size of arr to create, and q, the number of queries, respectively.
Each of the q subsequent lines contains a query string, queries[i].

Constraint
- 1<= n,q <= 10^5
- 0<= x,y <= 10^9
- It is guaranteed that query type 2 will never query an empty array or index.